requirements.txt
tensorflow>=2.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.30.0
matplotlib>=3.7.0
Pillow>=10.0.0
joblib>=1.3.0
